TCA-54 The IAuthDataSource and IRevokeTokenDataSource were unified
