+++ app/src/main/java/ar/scacchipa/twittercloneapp/data/datasource/IAuthExternalSource.kt
@@ -1,12 +1,13 @@
 package ar.scacchipa.twittercloneapp.data.datasource
 
+import ar.scacchipa.twittercloneapp.data.model.RevokeData
 import ar.scacchipa.twittercloneapp.data.model.UserAccessToken
 import retrofit2.Response
 import retrofit2.http.Field
 import retrofit2.http.FormUrlEncoded
 import retrofit2.http.POST
 
-interface IAuthDataSource {
+interface IAuthExternalSource {
     @FormUrlEncoded
     @POST("2/oauth2/token")
     suspend fun genAccessTokenSourceCode(
@@ -17,4 +18,12 @@ interface IAuthDataSource {
         @Field("code_verifier") codeVerifier: String,
         @Field("state") state:String
     ) : Response<UserAccessToken>
+
+    @FormUrlEncoded
+    @POST("2/oauth2/revoke")
+    suspend fun revokeToken(
+        @Field("token") token: String,
+        @Field("client_id") clientId: String,
+        @Field("token_type_hint") tokenTypeHint: String
+    ): Response<RevokeData>
 }
