+++ app/src/test/java/ar/scacchipa/twittercloneapp/data/repository/CredentialRepositoryTest.kt
@@ -1,8 +1,7 @@
 package ar.scacchipa.twittercloneapp.data.repository
 
 import ar.scacchipa.twittercloneapp.data.UserAccessTokenMapper
-import ar.scacchipa.twittercloneapp.data.datasource.IAuthDataSource
-import ar.scacchipa.twittercloneapp.data.datasource.IRevokeTokenDataSource
+import ar.scacchipa.twittercloneapp.data.datasource.IAuthExternalSource
 import ar.scacchipa.twittercloneapp.data.datasource.MockLocalStorage
 import ar.scacchipa.twittercloneapp.data.model.RevokeData
 import ar.scacchipa.twittercloneapp.data.model.UserAccessToken
@@ -28,14 +27,12 @@ import retrofit2.Response
 class CredentialRepositoryTest {
 
     private var mockCredentialLocalSource = MockLocalStorage()
-    private val mockAccessTokenExternalSource = mockk<IAuthDataSource>()
-    private var mockRevokeTokenDataSource = mockk<IRevokeTokenDataSource>()
+    private val mockAuthExternalSource = mockk<IAuthExternalSource>()
     private val hashMap = mutableMapOf<String, String>()
 
     private val subject = CredentialRepository(
         credentialLocalSource = mockCredentialLocalSource,
-        accessTokenExternalSource = mockAccessTokenExternalSource,
-        revokeDataSource = mockRevokeTokenDataSource,
+        authExternalSource = mockAuthExternalSource,
         mapper = UserAccessTokenMapper(),
         dispatcherDefault = Dispatchers.Default,
         dispatcherIO = Dispatchers.Default
@@ -45,7 +42,7 @@ class CredentialRepositoryTest {
     fun setup() {
 
         coEvery {
-            mockAccessTokenExternalSource.genAccessTokenSourceCode(
+            mockAuthExternalSource.genAccessTokenSourceCode(
                 transitoryToken = MockTokenProvider.transitoryToken1(),
                 grantType = "authorization_code",
                 clientId = "Yzg1a01Hcm16RTdKdmptZmhJdEs6MTpjaQ",
@@ -65,7 +62,7 @@ class CredentialRepositoryTest {
         )
 
         coEvery {
-            mockAccessTokenExternalSource.genAccessTokenSourceCode(
+            mockAuthExternalSource.genAccessTokenSourceCode(
                 "incorrect_password",
                 grantType = any(),
                 clientId = any(),
@@ -81,7 +78,7 @@ class CredentialRepositoryTest {
         )
 
         coEvery {
-            mockAccessTokenExternalSource.genAccessTokenSourceCode(
+            mockAuthExternalSource.genAccessTokenSourceCode(
                 transitoryToken = "request_without_body",
                 grantType = any(),
                 clientId = any(),
@@ -92,7 +89,7 @@ class CredentialRepositoryTest {
         } returns Response.success(null)
 
         coEvery {
-            mockAccessTokenExternalSource.genAccessTokenSourceCode(
+            mockAuthExternalSource.genAccessTokenSourceCode(
                 transitoryToken = "throw_exception",
                 grantType = any(),
                 clientId = any(),
@@ -227,7 +224,7 @@ class CredentialRepositoryTest {
         )
 
         coEvery {
-            mockRevokeTokenDataSource.revokeToken(
+            mockAuthExternalSource.revokeToken(
                 token = MockTokenProvider.credential1().accessToken,
                 clientId = Constants.CLIENT_ID,
                 tokenTypeHint = Constants.ACCESS_TOKEN
@@ -239,7 +236,7 @@ class CredentialRepositoryTest {
         subject.revokeCredential()
 
         coVerify {
-            mockRevokeTokenDataSource.revokeToken(
+            mockAuthExternalSource.revokeToken(
                 token = MockTokenProvider.credential1().accessToken,
                 clientId = Constants.CLIENT_ID,
                 tokenTypeHint = Constants.ACCESS_TOKEN
@@ -260,7 +257,7 @@ class CredentialRepositoryTest {
         )
 
         coEvery {
-            mockRevokeTokenDataSource.revokeToken(
+            mockAuthExternalSource.revokeToken(
                 token = MockTokenProvider.credential1().accessToken,
                 clientId = Constants.CLIENT_ID,
                 tokenTypeHint = Constants.ACCESS_TOKEN
@@ -272,7 +269,7 @@ class CredentialRepositoryTest {
         subject.revokeCredential()
 
         coVerify {
-            mockRevokeTokenDataSource.revokeToken(
+            mockAuthExternalSource.revokeToken(
                 token = MockTokenProvider.credential1().accessToken,
                 clientId = Constants.CLIENT_ID,
                 tokenTypeHint = Constants.ACCESS_TOKEN
